requirements.txt
streamlit==1.28.0
Pillow==10.3.0
streamlit-autorefresh==1.0.1



